+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/debugger/RobotValueGroup.java
@@ -12,14 +12,12 @@ import org.jetbrains.annotations.NotNull;
 public class RobotValueGroup extends XValueGroup {
 
     private final VariablesResponse variables;
-    private final int variableRef;
     private final IDebugProtocolServer debugServer;
     private final XDebugSession session;
 
-    public RobotValueGroup(String groupName, VariablesResponse variables, int variableRef, IDebugProtocolServer debugServer, XDebugSession session) {
+    public RobotValueGroup(String groupName, VariablesResponse variables, IDebugProtocolServer debugServer, XDebugSession session) {
         super(groupName);
         this.variables = variables;
-        this.variableRef = variableRef;
         this.debugServer = debugServer;
         this.session = session;
     }