+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/compilation/RobotImportAnnotator.java
@@ -1,12 +1,14 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.ide.inspections.compilation;
 
-import dev.xeonkryptos.xeonrobotframeworkplugin.RobotBundle;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Import;
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.Annotator;
 import com.intellij.lang.annotation.HighlightSeverity;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiReference;
+import dev.xeonkryptos.xeonrobotframeworkplugin.RobotBundle;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotHighlighter;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Import;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Variable;
 import org.jetbrains.annotations.NotNull;
 
 public class RobotImportAnnotator implements Annotator {
@@ -18,11 +20,15 @@ public class RobotImportAnnotator implements Annotator {
         }
 
         PsiElement[] children = importElement.getChildren();
-        if (children.length != 0 && children[0] == element) {
+        if (children.length != 0) {
             PsiElement child = children[0];
-            PsiReference reference = child.getReference();
-            if (reference == null || reference.resolve() == null) {
-                holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.getMessage("annotation.import.not-found")).range(child).create();
+            if (!(child.getFirstChild() instanceof Variable)) {
+                PsiReference reference = child.getReference();
+                if (reference == null || reference.resolve() == null) {
+                    holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.getMessage("annotation.import.not-found")).range(child).create();
+                } else {
+                    holder.newSilentAnnotation(HighlightSeverity.TEXT_ATTRIBUTES).textAttributes(RobotHighlighter.IMPORT_ARGUMENT).range(child).create();
+                }
             }
         }
     }