+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed not showing of Bracket settings in code completion
 - Automatically add ellipsis and indentation on enter in keyword statements which are preceded by a variable definition
 - Fixed too many started threads when a lot of file changes are detected at once
+- Annotation of import definitions that can't be resolved
 
 ### Changed